+++ e2e/test/scenarios/native-filters/sql-filters.cy.spec.js
@@ -78,7 +78,7 @@ describe("scenarios > filters > sql filters > basic filter types", () => {
         SQLFilter.toggleRequired();
         filterWidget().within(() => {
           cy.get("input").type("abc").should("have.value", "defaultabc");
-          cy.icon("time_history").click();
+          cy.icon("revert").click();
           cy.get("input").should("have.value", "default");
         });
       });
@@ -148,7 +148,7 @@ describe("scenarios > filters > sql filters > basic filter types", () => {
         SQLFilter.toggleRequired();
         filterWidget().within(() => {
           cy.get("input").type(".11").should("have.value", "3.11");
-          cy.icon("time_history").click();
+          cy.icon("revert").click();
           cy.get("input").should("have.value", "3");
         });
       });
@@ -234,7 +234,7 @@ describe("scenarios > filters > sql filters > basic filter types", () => {
           cy.findByText("15").click();
           cy.findByText("Update filter").click();
         });
-        filterWidget().icon("time_history").click();
+        filterWidget().icon("revert").click();
         filterWidget()
           .findByTestId("field-set-content")
           .should("have.text", "November 1, 2023");