+++ frontend/test/__support__/mocks.js
@@ -10,7 +10,9 @@ global.window.matchMedia = () => ({
 
 global.window.ResizeObserver = class ResizeObserver {
   observe() {}
+
   unobserve() {}
+
   disconnect() {}
 };