+++ src/metabase/query_analysis/native_query_analyzer.clj
@@ -31,7 +31,8 @@
 (def ^:private field-and-table-fragment
   "HoneySQL fragment to get the Field and Table"
   {:select    [[:f.id :field-id] [:f.name :column]
-               [:t.id :table-id] [:t.name :table]]
+               [:t.id :table-id] [:t.name :table]
+               [:t.schema :schema]]
    ;; (t2/table-name :model/Table) doesn't work on CI since models/table.clj hasn't been loaded
    :from      [[:metabase_table :t]]
    :left-join [[:metabase_field :f] [:= :f.table_id :t.id]]})
@@ -99,13 +100,16 @@
 
 (defn table-reference
   "Used by tests"
-  [db-id table]
-  (t2/select-one :model/QueryTable
-                 {:select [[:t.id :table-id] [:t.name :table]]
-                  :from   [[(t2/table-name :model/Table) :t]]
-                  :where  [:and
-                           [:= :t.db_id db-id]
-                           (table-query {:table (name table)})]}))
+  ([db-id table]
+   (table-reference db-id nil table))
+  ([db-id schema table]
+   (t2/select-one :model/QueryTable
+                  {:select [[:t.id :table-id] [:t.name :table] [:t.schema :schema]]
+                   :from   [[(t2/table-name :model/Table) :t]]
+                   :where  [:and
+                            [:= :t.db_id db-id]
+                            (table-query {:schema (some-> schema name)
+                                          :table (name table)})]})))
 
 (defn field-reference
   "Used by tests"
@@ -136,18 +140,31 @@
           references)))
 
 (defn- consolidate-columns
-  "Qualify analyzed columns with the corresponding database IDs, where we are able to resolve them."
+  "Qualify analyzed columns with the corresponding tables, where we are able to resolve them."
   [analyzed-columns database-columns]
-  ;; TODO we should handle schema as well
-  (let [->tab-key             (comp u/lower-case-en :table)
-        ->col-key             (comp u/lower-case-en :column)
-        column->records       (group-by ->col-key database-columns)
-        table+column->records (group-by (juxt ->tab-key ->col-key) database-columns)]
+  (let [->schema-key   (comp u/lower-case-en :schema)
+        ->table-key    (comp u/lower-case-en :table)
+        ->column-key   (comp u/lower-case-en :column)
+        ;; it may turn out cheaper to do scans versus building all these maps...
+        ;; alternately, a c->t->s trie could also make more sense
+        c->records     (group-by ->column-key database-columns)
+        t+c->records   (group-by (juxt ->table-key ->column-key) database-columns)
+        s+t+c->records (group-by (juxt ->schema-key ->table-key ->column-key) database-columns)]
     (strip-redundant-refs
-     (mapcat (fn [{:keys [table column] :as reference}]
-               (or (if table
-                     (table+column->records [(normalized-key table) (normalized-key column)])
-                     (column->records (normalized-key column)))
+     (mapcat (fn [{:keys [schema table column] :as reference}]
+               ;; match on what we know - where ambiguous, match with everything plausible
+               (or (cond
+                     schema
+                     (s+t+c->records [(normalized-key schema)
+                                      (normalized-key table)
+                                      (normalized-key column)])
+
+                     table
+                     (t+c->records [(normalized-key table)
+                                    (normalized-key column)])
+                     :else
+                     (c->records (normalized-key column)))
+                   ;; not found in the database records
                    [(update-vals reference strip-quotes)]))
              analyzed-columns))))
 
@@ -171,7 +188,7 @@
     (consolidate-tables
      tables
      (t2/select :model/QueryTable
-                {:select [[:t.id :table-id] [:t.name :table]]
+                {:select [[:t.id :table-id] [:t.name :table] [:t.schema :schema]]
                  :from   [[(t2/table-name :model/Table) :t]]
                  :where  [:and
                          [:= :t.db_id db-id]
@@ -188,8 +205,11 @@
                                      table-refs)
         merge-table-refs (fn [{:keys [schema table] :as field-ref}]
                            (map #(merge field-ref %)
-                                (if schema
+                                (cond
+                                  schema
                                   [(s+t->id [(normalize schema) (normalize table)])]
+
+                                  table
                                   (t->ids (normalize table)))))]
     (into (empty field-refs)
           (mapcat (fn [{:keys [table-id] :as field-ref}]